Inflation is now a lot more accurate

+++ firecasting/common-simulation/src/main/java/dk/gormkrings/factory/ISpecificationFactory.java
@@ -5,6 +5,6 @@ import dk.gormkrings.tax.ITaxRule;
 
 public interface ISpecificationFactory {
     ISpecification create(long startTime, ITaxRule taxRule);
-    ISpecification create(long startTime, ITaxRule taxRule, float inflation);
+    ISpecification create(long startTime, ITaxRule taxRule, double inflation);
 
 }
